v1: Brooks PF400: move DataID enum to its own module
Pure move of the parameter-database DataID enum into pylabrobot/brooks/data_ids.py;
precise_flex.py imports it. No behaviour change; gives one place to add more IDs.
